Changed logic for helper and fixed sung in url

+++ virtance/tasks.py
@@ -99,7 +99,8 @@ def action_virtance(virtance_id, action):
 def delete_virtance(virtance_id):
     virtance = Virtance.objects.get(id=virtance_id)
     wvcomp = WebVirtCompute(virtance.compute.token, virtance.compute.hostname)
-    wvcomp.delete_virtance(virtance.id)
-    ipaddresse = IPAddress.objects.filter(virtance=virtance)
-    ipaddresse.delete()
-    virtance.delete()
+    result = wvcomp.delete_virtance(virtance.id)
+    if result is True:
+        ipaddresse = IPAddress.objects.filter(virtance=virtance)
+        ipaddresse.delete()
+        virtance.delete()
